Create Blood Pressure Tracking (Simplified)
This code is based on data from 50 different customers and their blood pressure (BP). Using this code, I extracted the user's ID, average systolic and diastolic values, the date of entry, and the number of weeks elapsed since the initial entry. I had discovered a more simplified version of the first code, so here's the update.